feat: added worker threads for notifications
