+++ api/hub/package.json
@@ -64,6 +64,7 @@
     "node-abort-controller": "3.0.0",
     "object-hash": "2.2.0",
     "prom-client": "13.2.0",
+    "ts-node": "^10.2.1",
     "winston": "3.3.3"
   },
   "devDependencies": {
